♻️ GUI: Refactor GUI and CameraManager to use DLLoader for dynamic library loading
- Updated CameraManager to set mouse position using float values for better precision.
- Refactored GUI to utilize DLLoader for loading the RayLib dynamic library, improving modularity.
- Changed mouse position handling in GUI and HUD components to use Vector2i instead of std::pair<int, int>.
- Updated color definitions to use new color macros (CRED, CLIGHTGRAY, etc.) for consistency.
- Modified IDisplay interface to accommodate new mouse handling methods and color types.
- Implemented new DLLoader class for dynamic library management, including error handling.
- Adjusted various HUD and UI elements to ensure compatibility with the new color and display handling.
- Enhanced Raylib class to support new methods for mouse and camera interactions.

+++ gui/src/Graphic/Camera/CameraManager.cpp
@@ -79,7 +79,7 @@ void CameraManager::updateCameraTargetMode()
         int screenCenterX = screenSize.x / 2;
         int screenCenterY = screenSize.y / 2;
 
-        this->_display->setMousePosition({screenCenterX, screenCenterY});
+        this->_display->setMousePosition({static_cast<float>(screenCenterX), static_cast<float>(screenCenterY)});
         this->_display->disableCursor();
     }
 
@@ -103,7 +103,7 @@ void CameraManager::updateCameraTargetMode()
             Vector2i screenSize = this->_display->getScreenSize();
             int screenCenterX = screenSize.x / 2;
             int screenCenterY = screenSize.y / 2;
-            this->_display->setMousePosition({screenCenterX, screenCenterY});
+            this->_display->setMousePosition({static_cast<float>(screenCenterX), static_cast<float>(screenCenterY)});
         }
     }
 
@@ -191,7 +191,7 @@ void CameraManager::handlePlayerCameraMouseInput()
 
     if (this->_display->isMouseButtonPressed(MOUSE_BUTTON_RIGHT)) {
         _isPlayerViewDragging = true;
-        this->_display->setMousePosition({screenCenterX, screenCenterY});
+        this->_display->setMousePosition({static_cast<float>(screenCenterX), static_cast<float>(screenCenterY)});
         this->_display->disableCursor();
     }
 
@@ -205,7 +205,7 @@ void CameraManager::handlePlayerCameraMouseInput()
         const float rotationSensitivity = zappy::gui::CAMERA_SENSITIVITY;
 
         _playerAngleXZ += mouseDelta.x * rotationSensitivity;
-        this->_display->setMousePosition({screenCenterX, screenCenterY});
+        this->_display->setMousePosition({static_cast<float>(screenCenterX), static_cast<float>(screenCenterY)});
     }
 }
